Moved/refactored tests
